Refactored HTML reports to allow for reuse in other tools
